Avoid conflict with jQuery different versions
Some fetchers using jsdom module injects new versions of jQuery via the
'scripts' parameter of jsdom.env function. Sometimes, the pages loaded
already a specific version of the jQuerey injected and this can generate
conflicts on access to the $ object. The noConflict function must be
used to load the $ object.